Renamed mmr items to mrr
