+++ .bowerrc
@@ -1,3 +1,3 @@
 {
-  "directory": "mmr_website/static/libs"
+  "directory": "mrr_website/static/libs"
 }
