+++ Procfile
@@ -1,1 +1,1 @@
-web: gunicorn mmr-website.app:create_app\(\) -b 0.0.0.0:$PORT -w 3
+web: gunicorn mrr-website.app:create_app\(\) -b 0.0.0.0:$PORT -w 3
