+++ autoapp.py
@@ -2,7 +2,7 @@
 """Create an application instance."""
 from flask.helpers import get_debug_flag
 
-from mmr_website.app import create_app
-from mmr_website.settings import Config
+from mrr_website.app import create_app
+from mrr_website.settings import Config
 
 app = create_app(Config)
