+++ bower.json
@@ -1,5 +1,5 @@
 {
-  "name": "mmr-website",
+  "name": "mrr-website",
   "dependencies": {
     "bootstrap": "~3.2.0",
     "jQuery": "~2.1.3"
