+++ mrr_website/public/forms.py
@@ -4,7 +4,7 @@ from flask_wtf import Form
 from wtforms import PasswordField, StringField
 from wtforms.validators import DataRequired
 
-from mmr_website.user.models import User
+from mrr_website.user.models import User
 
 
 class LoginForm(Form):
