+++ mrr_website/static/js/common.js
