+++ mrr_website/templates/layout.html
@@ -10,7 +10,7 @@
 
   <title>{% block page_title %}
   
-    MMR Website
+    MRR Website
   
     {% endblock %}
   </title>
