+++ mrr_website/templates/nav.html
@@ -12,7 +12,7 @@
     </button>
     <a class="navbar-brand" href="{{ url_for('public.home') }}">
 
-      MMR Website
+      MRR Website
 
     </a>
   </div>
